AssetManager class implemented with createprojectile function!

+++ ECS/TransformComponent.h
@@ -11,7 +11,7 @@ public:
 
     int height = 32;
     int width = 32;
-    int scale = 1;
+    float scale = 1;
 
     int speed = 3;
 
@@ -31,7 +31,7 @@ public:
         position.Zero();
     }
 
-    TransformComponent(float x, float y, int h, int w, int sc)
+    TransformComponent(float x, float y, int h, int w, float sc)
     {
         position.x = x;
         position.y = y;
